requirements.txt
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
flake8==7.2.0
greenlet==3.2.3
idna==3.10
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
pynvim==0.5.2
python-dotenv==1.1.0
requests==2.32.4
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
